application/backend/requirements.txt
bcrypt==4.0.1
Django==3.2.6
django-cors-headers==3.7.0
PyMySQL==1.0.2
PyJWT==2.6.0
django-environ==0.10.0
cffi==1.15.1
pycparser==2.21
cryptography==39.0.2
uvicorn==0.17.6
gunicorn==20.1.0